# evidencije_laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Evidencije API",
        "description": "API za upravljanje predmetima, zadacima, predajama, upisima, kalendarom rokova i proverama plagijata.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/kalendar/rokovi": {
            "get": {
                "tags": [
                    "Kalendar"
                ],
                "summary": "Lista rokova zadataka i državnih praznika (Nager.Date API)",
                "operationId": "452f4fc5cdf619bbb0d3172aea0470b4",
                "responses": {
                    "200": {
                        "description": "Uspešno vraćeni rokovi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "zadatak-5"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "example": "internal"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Projektni zadatak"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Implementacija API-ja"
                                                    },
                                                    "start": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "end": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "all_day": {
                                                        "type": "boolean",
                                                        "example": false
                                                    },
                                                    "subject": {
                                                        "type": "string",
                                                        "example": "Internet tehnologije"
                                                    },
                                                    "subject_code": {
                                                        "type": "string",
                                                        "example": "ITEH"
                                                    },
                                                    "profesor": {
                                                        "type": "string",
                                                        "example": "Petar Petrović"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "external_calendar_provider": {
                                                    "type": "string",
                                                    "example": "nager_date_public_holidays"
                                                },
                                                "external_calendar_connected": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "today": {
                                                    "properties": {
                                                        "date": {
                                                            "type": "string",
                                                            "example": "2026-03-02"
                                                        },
                                                        "day_name": {
                                                            "type": "string",
                                                            "example": "ponedeljak"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Zabranjen pristup"
                    },
                    "401": {
                        "description": "Neautorizovan pristup"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}